mlr3fairness: Fairness Auditing and Debiasing for 'mlr3'

Integrates fairness auditing and bias mitigation methods for the 'mlr3' ecosystem. This includes fairness metrics, reporting tools, visualizations and bias mitigation techniques such as "Reweighing" described in 'Kamiran, Calders' (2012) doi:10.1007/s10115-011-0463-8 and "Equalized Odds" described in 'Hardt et al.' (2016) https://papers.nips.cc/paper/2016/file/9d2682367c3935defcb1f9e247a97c0d-Paper.pdf. Integration with 'mlr3' allows for auditing of ML models as well as convenient joint tuning of machine learning algorithms and debiasing methods.

See also

Author

Maintainer: Florian Pfisterer pfistererf@googlemail.com (ORCID)

Authors:

Wei Siyi weisiyi2@gmail.com

Michel Lang michellang@gmail.com (ORCID)