Classification Fair Generalized Ridge Regression Learner

Calls fairml::fgrrm from package fairml.

Details

Fair generalized ridge regression model implemented via package fairml. The 'unfairness' parameter is set to 0.05 as a default.

Dictionary

This Learner can be instantiated via the dictionary mlr_learners or with the associated sugar function lrn():

mlr_learners$get("classif.fairfgrrm")
lrn("classif.fairfgrrm")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”, “factor”, “ordered”

Required Packages: mlr3, fairml

Parameters

Id	Type	Default	Levels	Range

lambda	numeric	0		\([0, \infty)\)

definition	character	sp-komiyama	sp-komiyama, eo-komiyama	-

save.auxiliary	logical	FALSE	TRUE, FALSE	-

unfairness	numeric	-		\([0, 1]\)

family	character	binomial	gaussian, binomial	-

intersect	logical	TRUE	TRUE, FALSE	-

References

Scutari M, Panero F, Proissl M (2021). “Achieving Fairness with a Simple Ridge Penalty.” arXiv preprint arXiv:2105.13817.

See also

Author

pfistfl

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifFairfgrrm

Methods

Public methods

LearnerClassifFairfgrrm$new()

LearnerClassifFairfgrrm$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifFairfgrrm$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifFairfgrrm$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library("mlr3")
# stop example failing with warning if package not installed
learner = suppressWarnings(mlr3::lrn("classif.fairfgrrm"))
print(learner)
#> <LearnerClassifFairfgrrm:classif.fairfgrrm>
#> * Model: -
#> * Parameters: unfairness=0.05, intersect=FALSE
#> * Packages: mlr3, fairml
#> * Predict Types:  [response], prob
#> * Feature Types: integer, numeric, factor, ordered
#> * Properties: twoclass

# available parameters:
learner$param_set$ids()
#> [1] "lambda"         "definition"     "save.auxiliary" "unfairness"    
#> [5] "family"         "intersect"