Regression Fair Regression With Covariance Constraints Learner

Calls fairml::zlm from package fairml.

Details

Fair regression model from Zafar et al., 2019 implemented via package fairml. The 'unfairness' parameter is set to 0.05 as a default. The optimized fairness metric is statistical parity.

Dictionary

This Learner can be instantiated via the dictionary mlr_learners or with the associated sugar function lrn():

mlr_learners$get("regr.fairzlm")
lrn("regr.fairzlm")

Meta Information

Task type: “regr”

Predict Types: “response”

Feature Types: “integer”, “numeric”, “factor”, “ordered”

Required Packages: mlr3, fairml

Parameters

Id	Type	Default	Levels	Range

unfairness	numeric	-		\([0, 1]\)

intersect	logical	TRUE	TRUE, FALSE	-

References

BJ Z, I V, M G, KP G (2019). “Fairness Constraints: a Flexible Approach for Fair Classification.” In Journal of Machine Learning Research, 30, 1-42.

See also

Author

pfistfl

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrFairzlm

Methods

Public methods

LearnerRegrFairzlm$new()

LearnerRegrFairzlm$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrFairzlm$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrFairzlm$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library("mlr3")
# stop example failing with warning if package not installed
learner = suppressWarnings(mlr3::lrn("regr.fairzlm"))
print(learner)
#> <LearnerRegrFairzlm:regr.fairzlm>
#> * Model: -
#> * Parameters: unfairness=0.05, intersect=FALSE
#> * Packages: mlr3, fairml
#> * Predict Types:  [response]
#> * Feature Types: integer, numeric, factor, ordered
#> * Properties: -

# available parameters:
learner$param_set$ids()
#> [1] "unfairness" "intersect"